Django/requirements.txt
Django==1.5.4
South==0.8.2
djangorestframework==2.3.8
psycopg2==2.5.1
wsgiref==0.1.2
